+++ R/plot_metrics_ecdf.R
@@ -47,7 +47,7 @@
 #'
 plot_metrics_ecdf <- function(object, metrics = c("size", "sum_ys", "ratio_ys"),
                               restrict = NULL, nrow = NULL, ncol = NULL,
-                              cores = getOption("mc.cores", 1)) {
+                              cores = NULL) {
 
   # Check packages
   assert_pkgs("ggplot2")